requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.84
langchain-mistralai==0.2.12
langsmith==0.4.37
markdown-it-py==3.0.0
mdurl==0.1.2
orjson==3.11.5
packaging==25.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==15.0.0
shellingham==1.5.4
tavily-python==0.7.23
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
zstandard==0.25.0
